Move HKC entity classes into entity.py
