+++ homeassistant/components/homekit_controller/fan.py
@@ -20,7 +20,8 @@ from homeassistant.util.percentage import (
     ranged_value_to_percentage,
 )
 
-from . import KNOWN_DEVICES, HomeKitEntity
+from . import KNOWN_DEVICES
+from .entity import HomeKitEntity
 
 # 0 is clockwise, 1 is counter-clockwise. The match to forward and reverse is so that
 # its consistent with homeassistant.components.homekit.
