+++ homeassistant/components/homekit_controller/lock.py
@@ -17,7 +17,8 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import KNOWN_DEVICES, HomeKitEntity
+from . import KNOWN_DEVICES
+from .entity import HomeKitEntity
 
 CURRENT_STATE_MAP = {
     0: STATE_UNLOCKED,
