+++ homeassistant/components/homekit_controller/media_player.py
@@ -28,7 +28,8 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import KNOWN_DEVICES, HomeKitEntity
+from . import KNOWN_DEVICES
+from .entity import HomeKitEntity
 
 _LOGGER = logging.getLogger(__name__)
